feat: Add 3D model component for Estufa
The code changes include adding a new component called EstufaModel.js that renders a 3D model of a bread slice using the @react-three/drei and @react-spring/three libraries. This component is imported and used in the Estufa.js file to display the 3D model in the EstufaComponent. This feature enhances the visual representation of the Estufa and provides a more immersive user experience.